use cobra for cli
reorganize a bit to be able to use it
